objects/draw: merge Object_DrawPickupItem in TRX
