--- src/tr1/game/objects/common.h
@@ -1,5 +1,0 @@
-#pragma once
-
-// Generic collision and draw routines reused between various objects
-
-#include <libtrx/game/objects/types.h>
