--- src/tr2/game/objects/common.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <libtrx/game/objects/common.h>
-
-BOUNDS_16 Object_GetBoundingBox(
-    const OBJECT *obj, const ANIM_FRAME *frame, uint32_t mesh_bits);
